packages/python/requirements-dev.txt
# Development dependencies
python-dotenv>=1.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
